[Enhancement] Mob Types
This adds new functionality to the plugin making it possible for users to choose whether to block only monsters or all mobs.

This also address an issue when creating centers.yml where the "y" key of a center block would be stored with the quotes causing the plugin to fail in protecting areas.